backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "alembic>=1.18.4",
    "argon2-cffi>=25.1.0",
    "fastapi[standard]>=0.135.2",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.9.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.48",
]
